created class user repository 👌
